miAPIJWT/requirements.txt
uvicorn[standart]
fastapi
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
